+++ src/main/java/egovframework/let/cop/bbs/service/impl/EgovBBSLoneMasterServiceImpl.java
@@ -3,6 +3,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import egovframework.com.cmm.EgovMessageSource;
 import egovframework.let.cop.bbs.domain.model.BoardMaster;
 import egovframework.let.cop.bbs.domain.model.BoardMasterVO;
 import egovframework.let.cop.bbs.domain.repository.BBSLoneMasterDAO;
@@ -11,9 +12,12 @@ import egovframework.let.cop.com.service.BoardUseInf;
 import egovframework.let.cop.com.service.impl.BBSUseInfoManageDAO;
 
 import org.egovframe.rte.fdl.cmmn.EgovAbstractServiceImpl;
+import org.egovframe.rte.fdl.cmmn.exception.BaseRuntimeException;
 import org.egovframe.rte.fdl.idgnr.EgovIdGnrService;
 
 import jakarta.annotation.Resource;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.stereotype.Service;
 
@@ -31,10 +35,13 @@ import org.springframework.stereotype.Service;
  *  -------    --------    ---------------------------
  *  2009.08.25  한성곤          최초 생성
  *  2011.08.31  JJY            경량환경 템플릿 커스터마이징버전 생성 
+ *  2026.06.23  이백행          [2026년 컨트리뷰션] DAO 반환값 추가
  *  
  *  </pre>
  */
 @Service("EgovBBSLoneMasterService")
+@RequiredArgsConstructor
+@Slf4j
 public class EgovBBSLoneMasterServiceImpl extends EgovAbstractServiceImpl implements EgovBBSLoneMasterService {
 
     @Resource(name = "BBSLoneMasterDAO")
@@ -45,19 +52,33 @@ public class EgovBBSLoneMasterServiceImpl extends EgovAbstractServiceImpl implem
 
     @Resource(name = "egovBBSMstrIdGnrService")
     private EgovIdGnrService idgenService;
-		
+
+    private final EgovMessageSource egovMessageSource;
+
     /**
      * 등록된 게시판 속성정보를 삭제한다.
      */
     public void deleteMaster(BoardMaster boardMaster) throws Exception {
-		masterDAO.deleteMaster(boardMaster);
+		int result = masterDAO.deleteMaster(boardMaster);
+		if (log.isDebugEnabled()) {
+			log.debug("result={}", result);
+		}
+		if (result == 0) {
+			throw new BaseRuntimeException(egovMessageSource.getMessage("fail.common.update"));
+		}
 		
 		BoardUseInf bdUseInf = new BoardUseInf();
 		
 		bdUseInf.setBbsId(boardMaster.getBbsId());
 		bdUseInf.setLastUpdusrId(boardMaster.getLastUpdusrId());
 		
-		bbsUseDAO.deleteBBSUseInfByBoardId(bdUseInf);
+		int result2 = bbsUseDAO.deleteBBSUseInfByBoardId(bdUseInf);
+		if (log.isDebugEnabled()) {
+			log.debug("result2={}", result2);
+		}
+		if (result2 == 0) {
+			throw new BaseRuntimeException(egovMessageSource.getMessage("fail.common.update"));
+		}
     }
 
     /**
@@ -81,7 +102,13 @@ public class EgovBBSLoneMasterServiceImpl extends EgovAbstractServiceImpl implem
 		bdUseInf.setFrstRegisterId(boardMaster.getFrstRegisterId());
 		bdUseInf.setUseAt("Y");
 	
-		bbsUseDAO.insertBBSUseInf(bdUseInf);
+		int result = bbsUseDAO.insertBBSUseInf(bdUseInf);
+		if (log.isDebugEnabled()) {
+			log.debug("result={}", result);
+		}
+		if (result == 0) {
+			throw new BaseRuntimeException(egovMessageSource.getMessage("fail.common.insert"));
+		}
 		    
 		return bbsId;
     }
@@ -112,6 +139,12 @@ public class EgovBBSLoneMasterServiceImpl extends EgovAbstractServiceImpl implem
      * 게시판 속성정보를 수정한다.
      */
     public void updateMaster(BoardMaster boardMaster) throws Exception {
-    	masterDAO.updateMaster(boardMaster);
+    	int result = masterDAO.updateMaster(boardMaster);
+		if (log.isDebugEnabled()) {
+			log.debug("result={}", result);
+		}
+		if (result == 0) {
+			throw new BaseRuntimeException(egovMessageSource.getMessage("fail.common.update"));
+		}
     }
 }
